splits test from the code
